functions/python/requirements.txt
firebase-functions
firebase-admin
openai-whisper
